19/05/2023

+++ Back End/JobBoardAPI/Controllers/LocationsController.cs
@@ -14,21 +14,28 @@ namespace JobBoardAPI.Controllers
 
         // GET: api/<LocationsController>
         [HttpGet]
-        [Authorize]
-        public IEnumerable<Location> Get()
+        [Authorize(Policy = "admin")]
+        public IEnumerable<Location> GetLocations()
         {
             return _dbContext.Locations;
         }
 
         // GET api/<LocationsController>/5
         [HttpGet("{id}")]
-        public Location Get(int id)
+        [Authorize(Policy = "admin")]
+        public IActionResult GetLocationById(int id)
         {
-            return _dbContext.Locations.FirstOrDefault(x => x.Id == id);
+            var location = _dbContext.Locations.FirstOrDefault(x => x.Id == id);
+            if(location == null)
+            {
+                return NotFound();
+            }
+            return Ok(location);
         }
 
         // POST api/<LocationsController>
         [HttpPost]
+        [Authorize(Policy = "admin")]
         public void Post([FromBody] Location location)
         {
             _dbContext.Locations.Add(location);
@@ -37,22 +44,34 @@ namespace JobBoardAPI.Controllers
 
         // PUT api/<LocationsController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Location value)
+        [Authorize(Policy = "admin")]
+        public IActionResult Put(int id, [FromBody] Location value)
         {
             var location = _dbContext.Locations.Find(id);
+            if(location == null)
+            {
+                return NotFound("No records found with this id + " +  id);
+            }
             location.City = value.City;
             location.State = value.State;
             location.Country = value.Country;
             _dbContext.SaveChanges();
+            return Ok("Record updated successfully");
         }
 
         // DELETE api/<LocationsController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        [Authorize(Policy = "admin")]
+        public IActionResult Delete(int id)
         {
             var location = _dbContext.Locations.Find(id);
+            if(location == null)
+            {
+                return NotFound("No records found with this id + " + id);
+            }
             _dbContext.Locations.Remove(location);
             _dbContext.SaveChanges();
+            return Ok("Record Deleted");
         }
     }
 }
